Extract `dyn_info.js`
